+++ frontend/openchat-agent/src/services/userIndex/userIndex.client.ts
@@ -67,7 +67,7 @@ import {
     UserIndexSetInternalModerationChannelArgs,
     UserIndexResolveModerationReportArgs,
     UserIndexSetModerationFlagsArgs,
-    UserIndexSetOpenAiApiKeyArgs,
+    UserIndexSetOpenaiApiKeyArgs,
     UserIndexSetPremiumItemCostArgs,
     UserIndexSetUsernameArgs,
     UserIndexSetUsernameResponse,
@@ -179,7 +179,7 @@ export class UserIndexClient extends SingleCanisterMsgpackAgent {
                 api_key: apiKey === undefined || apiKey === "" ? undefined : apiKey,
             },
             (resp) => resp === "Success",
-            UserIndexSetOpenAiApiKeyArgs,
+            UserIndexSetOpenaiApiKeyArgs,
             UnitResult,
         );
     }
@@ -204,15 +204,26 @@ export class UserIndexClient extends SingleCanisterMsgpackAgent {
         );
     }
 
-    resolveModerationReport(reportIndex: bigint, verdict: ModerationVerdict): Promise<boolean> {
+    resolveModerationReport(reportIndex: bigint, verdict: ModerationVerdict, urgent: boolean | undefined): Promise<boolean> {
         return this.update(
             "resolve_moderation_report",
             {
                 report_index: reportIndex,
                 verdict: apiModerationVerdict(verdict),
+                urgent,
             },
             (resp) => resp === "Success",
             UserIndexResolveModerationReportArgs,
+            UnitResult,
+        );
+    }
+
+    contestModerationSanction(): Promise<boolean> {
+        return this.update(
+            "contest_moderation_sanction",
+            {},
+            (resp) => resp === "Success",
+            Empty,
             UnitResult,
         );
     }